This will add the 5-rev_string.c file

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -11,7 +11,7 @@ void rev_string(char *s)
 {
 	int i = 0;
 	int k = 0;
-	int j;
+	long unsigned int j;
 
 	for (j = 0; j <= (strlen(s) - 1); j++)
 	{
